+++ data-r2dbc/src/main/java/com/example/demo/PostRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.r2dbc.repository.query.Query;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 
-interface PostRepository extends ReactiveCrudRepository<Post, Long> {
+interface PostRepository extends ReactiveCrudRepository<Post, Integer> {
 
     @Query("SELECT * FROM posts WHERE title like $1")
     Flux<Post> findByTitleContains(String name);
